notIn, requiredUnlessData validators: their camelCased naming is broken, let's rename them to-use-dashes-instead.
